FormHandlerExpressionEvaluationContextTesting2 extracted FormHandlerExpressionEvaluationContextTesting
