+++ src/main/java/walkingkooka/validation/form/expression/FormHandlerExpressionEvaluationContextTesting2.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2025 Miroslav Pokorny (github.com/mP1)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package walkingkooka.validation.form.expression;
+
+import walkingkooka.tree.expression.ExpressionEvaluationContextTesting2;
+import walkingkooka.validation.ValidationReference;
+import walkingkooka.validation.form.FormHandlerContextTesting2;
+
+public interface FormHandlerExpressionEvaluationContextTesting2<R extends ValidationReference, S, C extends FormHandlerExpressionEvaluationContext<R, S>> extends FormHandlerExpressionEvaluationContextTesting,
+    ExpressionEvaluationContextTesting2<C>,
+    FormHandlerContextTesting2<C, R, S> {
+
+    @Override
+    default C createConverterLike() {
+        return this.createContext();
+    }
+
+    @Override
+    default String typeNameSuffix() {
+        return FormHandlerExpressionEvaluationContext.class.getSimpleName();
+    }
+}
